Refactor Header.jsx to Header.js preventing dependency cycle.
